implemented absolute URL for application page
