+++ frontend/app/applications/ApplicationsBoard.js
