+++ frontend/app/applications/StatusColumn.js
