+++ frontend/app/applications/[id]/DashboardApplicationHeader.js
@@ -10,31 +10,8 @@ import {
 } from '@mui/material'
 import CancelRoundedIcon from '@mui/icons-material/CancelRounded'
 import DeleteRoundedIcon from '@mui/icons-material/DeleteRounded'
-import { useApplicationById } from '@/app/hooks/useApplicationById'
 
-export default function DashboardApplicationHeader() {
-  const params = useParams()
-  const id = params.id
-  const { application, isLoading, error } = useApplicationById(id)
-  if (isLoading) {
-    return (
-      <Box sx={{ display: 'flex', justifyContent: 'center', p: 4 }}>
-        <CircularProgress />
-      </Box>
-    )
-  }
-  if (error) {
-    return (
-      <Alert severity="error" sx={{ m: 4 }}>
-        {error}
-      </Alert>
-    )
-  }
-
-  if (!application) {
-    return <Typography>No application data available</Typography>
-  }
-
+export default function DashboardApplicationHeader({ application }) {
   return (
     <>
       <Stack
@@ -46,7 +23,7 @@ export default function DashboardApplicationHeader() {
             {application.job_title}
           </Typography>
           <Typography variant="h5" component="h5" sx={{ pb: 2 }}>
-            {application.name.toUpperCase()}
+            {application.name.toLocaleUpperCase()}
           </Typography>
         </Stack>
 
@@ -56,7 +33,7 @@ export default function DashboardApplicationHeader() {
           sx={{ justifyContent: 'space-between', alignItems: 'center' }}
         >
           <Typography sx={{ color: 'accent.main' }}>
-            {application.status.toUpperCase()}
+            {application.status.toLocaleUpperCase()}
           </Typography>
           <IconButton size="large" sx={{ color: 'accent.main' }}>
             <DeleteRoundedIcon />
