+++ frontend/app/applications/[id]/Notes.js
