+++ frontend/app/applications/[id]/Overview.js
@@ -1,8 +1,5 @@
 'use client'
-import { useParams } from 'next/navigation'
 import {
-  CircularProgress,
-  Alert,
   Box,
   Typography,
   Link,
@@ -15,32 +12,10 @@ import {
   useTheme,
   useMediaQuery,
 } from '@mui/material'
-import { useApplicationById } from '@/app/hooks/useApplicationById'
 
-const Overview = () => {
-  const params = useParams()
+const Overview = ({ application }) => {
   const theme = useTheme()
   const isMobile = useMediaQuery(theme.breakpoints.down('md'))
-  const id = params.id
-  const { application, isLoading, error } = useApplicationById(id)
-  if (isLoading) {
-    return (
-      <Box sx={{ display: 'flex', justifyContent: 'center', p: 4 }}>
-        <CircularProgress />
-      </Box>
-    )
-  }
-  if (error) {
-    return (
-      <Alert severity="error" sx={{ m: 4 }}>
-        {error}
-      </Alert>
-    )
-  }
-
-  if (!application) {
-    return <Typography>No application data available</Typography>
-  }
 
   return (
     <Box
