+++ frontend/app/hooks/useApplicationById.js
@@ -22,7 +22,7 @@ export function useApplicationById(id) {
       setError(
         err.status === 401
           ? 'Please login to view your application.'
-          : `We're having trouble loading your application with ${id}. Please try again later.`
+          : `We're having trouble loading your application with ID ${id}. Please try again later.`
       )
     } finally {
       setIsLoading(false)
