+++ frontend/app/user/page.js
@@ -8,7 +8,7 @@ import TabsControl from './tabs/TabsControl'
 import SearchField from './SearchField'
 import SortControl from './SortControl'
 import TabPanel from './tabs/TabPanel'
-import ApplicationsBoard from './applications/ApplicationsBoard'
+import ApplicationsBoard from '../applications/ApplicationsBoard'
 import MotivationalQuote from './MotivationalQuote'
 import { getLocalStorageItem } from '@/utils/localStorage'
 import { useApplications } from '../hooks/useApplications'
